fix segmenatation fault on 64 bit platforms.... CURL_INFO_PRIVATE should accomodate cchar* which correspond to 'long' on 64bit platforms

+++ libs/util/http/src/curl/CurlHttpRequest.cpp
@@ -405,7 +405,7 @@ int curlHTTPProgress(void * curlHttpRequestInstance, double dltotal, double dlno
 		CurlHttpRequest * instance = (CurlHttpRequest *) curlHttpRequestInstance;
 		instance->downloadDone = dlnow;
 		instance->downloadTotal = dltotal;
-		int requestId;
+		long requestId;
 		curl_easy_getinfo(instance->_curl, CURLINFO_PRIVATE, &requestId);
 		instance->dataReadProgressEvent(requestId, dlnow, dltotal);
 		instance->dataSendProgressEvent(requestId, ulnow, ultotal);
